# openapi.yaml
openapi: 3.0.3
info:
  title: Payment Orchestration Platform API
  description: |
    REST API for the Payment Orchestration Platform.

    All client-facing APIs are routed through the API Gateway.
    Authentication is handled using JWT bearer tokens.

    The platform provides:
    - User registration and authentication
    - Password management
    - User retrieval
    - Payment creation and retrieval
    - Gateway health monitoring
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local API Gateway
...
paths:
...
  /api/v1/payments:
    post:
      tags:
        - Payments
      summary: Create a payment
      operationId: createPayment

      security:
        - bearerAuth: []

      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: |
            Unique key used to safely retry payment requests
            without creating duplicate transactions.
          schema:
            type: string
          example: idem-ORD-1001-001

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
            example:
              amount: 1999
              currency: INR
              description: Test payment
              metadata:
                orderId: ORD-1001

      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'

        '400':
          description: Invalid payment request

        '401':
          description: Unauthorized

        '409':
          description: Duplicate idempotency key

        '502':
          description: Payment provider error
  /api/v1/payments/{transactionId}:
    get:
      tags:
        - Payments
      summary: Get payment by transaction ID
      operationId: getPaymentByTransactionId

      security:
        - bearerAuth: []

      parameters:
        - name: transactionId
          in: path
          required: true
          description: Unique payment transaction identifier
          schema:
            type: string
          example: txn_12345

      responses:
        '200':
          description: Payment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'

        '401':
          description: Unauthorized

        '404':
          description: Transaction not found
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the login endpoint.
  schemas:
...
    CreatePaymentRequest:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Payment amount in the smallest currency unit.
          example: 1999

        currency:
          type: string
          description: ISO 4217 currency code.
          example: INR

        description:
          type: string
          example: Test payment

        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            orderId: ORD-1001
    PaymentResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: txn_12345

        status:
          type: string
          example: SUCCESS

        amount:
          type: integer
          format: int64
          example: 1999

        currency:
          type: string
          example: INR

        description:
          type: string
          example: Test payment

        stripePaymentIntentId:
          type: string
          example: pi_123456789

        idempotencyKey:
          type: string
          example: idem-ORD-1001-001

        createdAt:
          type: string
          format: date-time
          example: 2026-08-29T12:30:00Z